Now it's possible to get the ability name for a unit's current command. That's also used for minimap destination coloring. Also, most address variables have been changed from signed to unsigned.

+++ Data_Source/Data/GameData.cs
@@ -172,7 +172,7 @@ namespace Data
 				map.mapInfo = new MapInfo();
 				return map;
 			}
-			byte[] data = offsets.ReadStruct(ORNames.MapFileInfo, (int)num);
+			byte[] data = offsets.ReadStruct(ORNames.MapFileInfo, num);
 
 			map_information_s _s = new map_information_s();
 			_s = (map_information_s) mem.ReadMemory(num, typeof(map_information_s));
